Fixed common path issues and organized functions
