+++ src/bin/rimage/paths/tests.rs
